Move slides-specific CI concerns to slides
CI config (& tasks) for a specific component should live in that component.
Some teams put it in a separate, central CI repo, but this goes against
the best practices.

For example, when a component changes the way its tests are run, or the
way deps are packaged, the CI config should be updated at the same time,
within the same repo.

Same for cases when a git resource version is linked to a specific
path(s). Changes in the `concourse` dir are not visible to the `app`
resource because it's only concerned with the `slides` path...